+++ src/index.js
@@ -1,5 +1,7 @@
 
 const {gitP} = require('./lib/runners/promise-wrapped');
-const {gitInstanceFactory, gitExportFactory} = require('./git-factory');
+const {esModuleFactory, gitInstanceFactory, gitExportFactory} = require('./git-factory');
 
-module.exports = gitExportFactory(gitInstanceFactory, {gitP});
+module.exports = esModuleFactory(
+   gitExportFactory(gitInstanceFactory, {gitP})
+);