+++ typings/response.d.ts
@@ -1,5 +1,62 @@
 import { DefaultLogFields } from '../src/lib/tasks/log';
 
+export interface BranchSummaryBranch {
+   current: boolean;
+   name: string;
+   commit: string;
+   label: string;
+}
+
+export interface BranchSummary {
+   detached: boolean;
+   current: string;
+   all: string[];
+   branches: {
+      [key: string]: BranchSummaryBranch;
+   };
+}
+
+/**
+ * Represents the status of a single branch deletion
+ */
+export interface BranchDeletionSummary {
+   branch: string;
+   hash: string | null;
+   readonly success: boolean;
+}
+
+/**
+ * Represents the status of having deleted a batch of branches
+ */
+export interface BranchDeletionBatchSummary {
+   /**
+    * All branches included in the response
+    */
+   all: BranchDeletionSummary[];
+
+   /**
+    * Branches mapped by their branch name
+    */
+   branches: { [branchName: string]: BranchDeletionSummary };
+
+   /**
+    * Array of responses that are in error
+    */
+   errors: BranchDeletionSummary[];
+
+   /**
+    * Flag showing whether all branches were deleted successfully
+    */
+   readonly success: boolean;
+}
+
+export interface CleanSummary {
+   readonly dryRun: boolean;
+   paths: string[];
+   files: string[];
+   folders: string[];
+}
+
 export interface CommitSummary {
    author: null | {
       email: string;
@@ -12,6 +69,36 @@ export interface CommitSummary {
       insertions: number;
       deletions: number;
    };
+}
+
+/**
+ * Represents the current git configuration, as defined by the output from `git log`
+ */
+export interface ConfigListSummary {
+
+   /**
+    * All configuration settings, where local/user settings override user/global settings
+    * the overridden value will appear in this object.
+    */
+   readonly all: ConfigValues;
+
+   /**
+    * The file paths configuration was read from
+    */
+   files: string[];
+
+   /**
+    * The `ConfigValues` for each of the `files`, use this object to determine
+    * local repo, user and global settings.
+    */
+   values: {[fileName: string]: ConfigValues};
+}
+
+/**
+ * Represents the map of configuration settings
+ */
+export interface ConfigValues {
+   [key: string]: string | string[];
 }
 
 export interface DiffResultTextFile {